Update translation.js
Complete translation.
Correct, "Universal Canning Machine" was originally named "Fluid / Solid Canning Machine"; "Liquid Fuel Firebox" The name of the machine is different from the GUI name. The former is American English and the latter is British English. The author is requested to see that these words can be corrected in time. Thanks!